Update contribute.rmd
This was my first time doing this sort of editing/pull via GitHub and I was a bit timid as I didn't find the options to "save" my changes in the editor.  Also, you're able to describe the change here at below the editor, prior to submitting the pull request, so I edited the directions to reflect that (if there is the option to provide a description at the pull request as well, which is preferable?).

A period was added to the end of the first bullet for standardization.